# crates/dagron-engine/openapi.yaml
openapi: 3.0.3
info:
  title: module-54 workflow scheduler — management API
  description: >
    Read/write control surface over the workflow scheduler's datastore: list and
    inspect runs, submit a DAG, cancel a run, and scrape metrics. Every operation
    is a thin shell over the same `db` facade the reconcile loop uses, so the API
    holds no scheduler state of its own and works whether it runs in-process next
    to the loop or as a standalone read sidecar against the shared Postgres
    backend.


    Served only when the binary is built with the (default) `ops` Cargo feature
    and `API_ADDR` is set. This document is also served live by the running
    scheduler at `/openapi.yaml` and `/openapi.json`, with a Swagger UI at `/docs`.


    Security: this embedded API performs **no authentication** of its own — bind
    `API_ADDR` to localhost / a trusted network and place authentication (e.g. the
    `dagron-api` service, an auth proxy, or network policy) in front of it before
    exposing it. It is intended for in-process ops and trusted-network sidecar use.
  version: 0.1.0
  license:
    name: lab crate
servers:
  - url: http://127.0.0.1:8080
    description: Local management API (set API_ADDR to choose host:port)
...
paths:
...
  /runs:
    get:
      tags: [runs]
      summary: List runs
      description: Runs newest-first, joined to the definition for the DAG name.
      operationId: listRuns
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by run status.
          schema: { $ref: "#/components/schemas/RunStatus" }
        - name: limit
          in: query
          required: false
          description: Max rows to return (clamped to 1..=1000).
          schema: { type: integer, format: int64, default: 50, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: Matching runs
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    maxItems: 1000
                    items: { $ref: "#/components/schemas/RunSummary" }
        "500":
          $ref: "#/components/responses/Error"
    post:
      tags: [runs]
      summary: Submit a DAG
      description: >
        Parse and cycle-check a DAG (YAML — a superset of JSON, so JSON bodies
        parse too) and create a run for it. The raw body is stored verbatim as the
        workflow definition.


        Admission control: when `MAX_INFLIGHT_RUNS` is set and that many runs are
        already active, the submit is shed with `429 Too Many Requests` and a
        `Retry-After` header instead of growing an unbounded backlog. Clients
        should back off and retry.
      operationId: submitRun
      parameters:
        - name: wait
          in: query
          required: false
          description: >
            Synchronous invocation: block until the run reaches a terminal state
            (or `timeout_secs` elapses) and return its status + result instead of
            just the id. Response is `200` (not `201`) in this mode.
          schema: { type: boolean, default: false }
        - name: timeout_secs
          in: query
          required: false
          description: Max seconds to wait when `wait=true` (default 30, clamped 1–600).
          schema: { type: integer, default: 30, minimum: 1, maximum: 600 }
      requestBody:
        required: true
        description: The DAG spec.
        content:
          application/yaml:
            schema: { $ref: "#/components/schemas/DagSpec" }
          application/json:
            schema: { $ref: "#/components/schemas/DagSpec" }
          text/plain:
            schema: { type: string }
      responses:
        "201":
          description: Run created (fire-and-forget; `wait` absent/false)
          content:
            application/json:
              schema:
                type: object
                required: [run_id]
                properties:
                  run_id: { type: string, format: uuid }
        "200":
          description: >
            Run created and awaited (`wait=true`). A timed-out wait still returns
            `200` with `finished: false` and the live status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunResult" }
        "400":
          description: Invalid DAG (parse error, unknown dependency, or cycle)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: >
            At the in-flight admission cap (`MAX_INFLIGHT_RUNS`). Back off and
            retry after the `Retry-After` header.
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
  /runs/{id}:
    get:
      tags: [runs]
      summary: Get a run and its tasks
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run detail
          content:
            application/json:
              schema:
                type: object
                required: [run, tasks]
                properties:
                  run: { $ref: "#/components/schemas/WorkflowRun" }
                  tasks:
                    type: array
                    items: { $ref: "#/components/schemas/TaskRun" }
        "404":
          description: No such run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      description: Run UUID.
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Unexpected server/datastore error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable failure message }
...
    RunStatus:
      type: string
      description: Workflow run lifecycle state.
      enum: [pending, running, succeeded, failed, cancelled]
    TaskStatus:
      type: string
      description: Task state-machine node.
      enum: [pending, ready, running, awaiting_approval, succeeded, failed, skipped, cancelled]
    RunSummary:
      type: object
      description: List-view projection of a run, with the DAG name joined in.
      required: [id, name, status, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, description: DAG name from the definition }
        status: { $ref: "#/components/schemas/RunStatus" }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
    WorkflowRun:
      type: object
      description: One execution instance of a workflow definition.
      required: [id, definition_id, status, created_at]
      properties:
        id: { type: string, format: uuid }
        definition_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/RunStatus" }
        input: { type: string, nullable: true }
        output: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
    RunResult:
      type: object
      description: >
        The outcome of a synchronous invocation (`POST /runs?wait=true` /
        `GET /runs/{id}/wait`). `result` is the `result_from` task's output on
        success, otherwise null; `finished` is false only when the wait timed out.
      required: [run_id, status, finished]
      properties:
        run_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/RunStatus" }
        finished:
          type: boolean
          description: True once the run is terminal; false if the wait timed out.
        result:
          type: string
          nullable: true
          description: The `result_from` task's output on success, else null.
...
    TaskRun:
      type: object
      description: One node within a run — the row that drives the state machine.
      required: [id, run_id, name, status, attempt, remaining_deps, version]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        name: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        attempt: { type: integer, format: int64, description: Attempts run so far }
        remaining_deps:
          type: integer
          format: int64
          description: Dependency counter; task becomes ready at 0
        input:
          type: string
          nullable: true
          description: Serialised TaskSpec JSON
        output: { type: string, nullable: true, description: stdout or error message }
        claimed_by: { type: string, nullable: true, description: Worker holding the lease }
        lease_expires_at: { type: string, format: date-time, nullable: true }
        version:
          type: integer
          format: int64
          description: Monotonic CAS/fencing token guarding against double-claim
        scheduled_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        pool: { type: string, nullable: true, description: Named concurrency pool this task draws a slot from }
        cache_hit:
          type: boolean
          description: True when the task was resolved from the memoization cache instead of executed
        wake_at:
          type: string
          format: date-time
          nullable: true
          description: >-
            Park reason — a `type: wait` time sensor resumes at this instant.
            A parked task stays `running` with a NULL lease, so these four
            fields (at most one set) are what distinguishes "waiting on
            something" from "hung".
        wait_url:
          type: string
          nullable: true
          description: Park reason — endpoint a `wait.url` sensor polls until it answers 2xx
        wait_dataset:
          type: string
          nullable: true
          description: Park reason — dataset URI a `wait.dataset` sensor waits to see updated
        sub_run_id:
          type: string
          format: uuid
          nullable: true
          description: "Park reason — child run a `type: workflow` trigger is blocked on"
    TaskSpec:
      type: object
      description: One task in a submitted DAG.
      required: [name, command]
      properties:
        name: { type: string }
        command:
          type: array
          items: { type: string }
          description: argv; command[0] is the binary
          example: ["echo", "hello"]
        depends_on:
          type: array
          items: { type: string }
          default: []
          description: Names of tasks that must succeed first
        input:
          nullable: true
          description: Arbitrary JSON payload carried with the task
        max_attempts:
          type: integer
          format: int32
          default: 1
          minimum: 1
          description: Attempt cap before the task is marked failed (1 = no retries)
        retry_delay_secs:
          type: integer
          format: int64
          default: 0
          description: Base backoff; actual delay = retry_delay_secs * 2^(attempt-1)
        timeout_secs:
          type: integer
          format: int64
          nullable: true
          description: Per-task timeout (falls back to 25 s when absent)
        docker_image:
          type: string
          nullable: true
          description: Image for the Docker/Kubernetes executor; ignored by LocalExecutor
        env:
          type: array
          default: []
          description: Environment variables injected into the task container
          items:
            type: object
            required: [name, value]
            properties:
              name: { type: string }
              value: { type: string }
        resources:
          nullable: true
          description: >
            Per-task pod CPU/memory requests & limits (Kubernetes executor only).
          type: object
          properties:
            requests:
              type: object
              additionalProperties: { type: string }
              example: { cpu: "250m", memory: "256Mi" }
            limits:
              type: object
              additionalProperties: { type: string }
              example: { memory: "512Mi" }
        service_account:
          type: string
          nullable: true
          description: >
            ServiceAccount for the task pod — the IRSA seam for S3 access
            (Kubernetes executor only).
    DagSpec:
      type: object
      description: A submitted workflow DAG.
      required: [name, tasks]
      properties:
        name: { type: string }
        result_from:
          type: string
          nullable: true
          description: >
            Name of the task whose output becomes the run's result (returned by a
            synchronous invocation). Must name a real, non-hook task.
        tasks:
          type: array
          items: { $ref: "#/components/schemas/TaskSpec" }
      example:
        name: simple-pipeline
        tasks:
          - { name: prepare, command: ["echo", "step 1"], depends_on: [] }
          - { name: summarize, command: ["echo", "done"], depends_on: [prepare] }